+++ app/javascript/Components/Result/flexible_criterion_input.jsx
@@ -146,7 +146,7 @@ export default function FlexibleCriterionInput({
 
   useEffect(() => {
     if (active) {
-      liRef.current?.scrollIntoView?.({block: "nearest", behavior: "smooth"});
+      liRef.current?.scrollIntoView({block: "nearest", behavior: "smooth"});
     }
   }, [active]);